changed Double to BigDecimal

+++ src/main/java/all/that/matters/model/Food.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -27,7 +28,7 @@ public class Food {
     private String name;
 
     @Column(name = "calories", nullable = false)
-    private Double calories;
+    private BigDecimal calories;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id")
